Refactor error parser test service into trait
