+++ tests/Api/ErrorParser/RestJsonErrorParserTest.php
@@ -2,18 +2,17 @@
 namespace Aws\Test\Api\ErrorParser;
 
 use Aws\Api\ErrorParser\RestJsonErrorParser;
-use Aws\Api\Parser\Exception\ParserException;
-use Aws\Api\Service;
-use Aws\AwsClient;
 use GuzzleHttp\Psr7;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers Aws\Api\ErrorParser\RestJsonErrorParser
- * @covers Aws\Api\ErrorParser\JsonParserTrait
+ * @covers \Aws\Api\ErrorParser\RestJsonErrorParser
+ * @covers \Aws\Api\ErrorParser\JsonParserTrait
  */
 class RestJsonErrorParserTest extends TestCase
 {
+    use ErrorParserTestServiceTrait;
+
     /**
      * @dataProvider errorResponsesProvider
      *
@@ -37,7 +36,7 @@ class RestJsonErrorParserTest extends TestCase
 
     public function errorResponsesProvider()
     {
-        $service = $this->generateTestService();
+        $service = $this->generateTestService('rest-json');
         $client = $this->generateTestClient($service);
         $command = $client->getCommand('TestOperation', []);
 
@@ -189,91 +188,5 @@ class RestJsonErrorParserTest extends TestCase
                 ]
             ],
         ];
-    }
-
-    private function generateTestClient(Service $service, $args = [])
-    {
-        return new AwsClient(
-            array_merge(
-                [
-                    'service'      => 'foo',
-                    'api_provider' => function () use ($service) {
-                        return $service->toArray();
-                    },
-                    'region'       => 'us-east-1',
-                    'version'      => 'latest',
-                ],
-                $args
-            )
-        );
-    }
-
-    private function generateTestService()
-    {
-        return new Service(
-            [
-                'metadata' => [
-                    "protocol" => 'rest-json',
-                    "apiVersion" => "2019-05-01"
-                ],
-                'shapes' => [
-                    "HeaderMap"=> [
-                        "type"=> "map",
-                        "key"=> [
-                            "shape"=> "String"
-                        ],
-                        "value"=> [
-                            "shape"=> "String"
-                        ]
-                    ],
-                    "Integer" => ["type" => "integer"],
-                    "String" => ["type" => "string"],
-                    "TestException"=>[
-                        "type" => "structure",
-                        "members" => [
-                            "TestString" => ["shape" => "String"],
-                            "TestInt" => ["shape" => "Integer"],
-                            "TestHeaderMember" => [
-                                "shape" => "String",
-                                "location" => "header",
-                                "locationName" => "TestHeader",
-                            ],
-                            "TestHeaders" => [
-                                "shape" => "HeaderMap",
-                                "location" => "headers",
-                                "locationName" => "x-meta-",
-                            ],
-                            "TestStatus" => [
-                                "shape" => "Integer",
-                                "location" => "statusCode",
-                            ],
-                        ],
-                        "error" => ["httpStatusCode" => 502],
-                        "exception" => true,
-                    ],
-                    "TestInput"=>[
-                        "type" => "structure",
-                        "members" => [
-                            "TestInput" => ["shape" => "String"]
-                        ],
-                    ],
-                ],
-                'operations' => [
-                    "TestOperation"=> [
-                        "name"=> "TestOperation",
-                        "http"=> [
-                            "method"=> "POST",
-                            "requestUri"=> "/",
-                            "responseCode"=> 200
-                        ],
-                        "input" => ["shape"=> "TestInput"],
-                        "errors" => [
-                            ["shape" => "TestException"]
-                        ],
-                    ],
-                ],
-            ],
-            function () { return []; }
-        );
     }
 }
